fix: move parse to util
