+++ src/screens/QrScanner.js
@@ -16,7 +16,6 @@
 
 'use strict';
 
-import { encodeAddress } from '@polkadot/util-crypto';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { Alert, StyleSheet, Text, View } from 'react-native';
@@ -48,11 +47,15 @@ export default class Scanner extends React.PureComponent {
               navigation={this.props.navigation}
               scannerStore={scannerStore}
               onBarCodeRead={async txRequestData => {
+                debugger;
                 if (scannerStore.isBusy()) {
                   return;
                 }
 
+                debugger;
+
                 try {
+                  debugger;
                   scannerStore.parseRawData(txRequestData.rawData);
                 } catch (e) {
                   Alert.alert('Unable to parse transaction', e.message, [
