+++ src/util/rawDataToU8A.js
@@ -1,0 +1,80 @@
+// Copyright 2015-2019 Parity Technologies (UK) Ltd.
+// This file is part of Parity.
+
+// Parity is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Parity is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
+
+// @flow
+
+/*
+  Example Full Raw Data
+  ---
+  4 // indicates binary
+  37 // indicates data length
+  0000 // frame count
+  0100 // first frame
+  --- UOS Specific Data
+  53 // indicates payload is for Substrate
+  01 // crypto: sr25519
+  00 // indicates action: signData
+  f4cd755672a8f9542ca9da4fbf2182e79135d94304002e6a09ffc96fef6e6c4c // public key
+  544849532049532053504152544121 // actual payload message to sign (should be SCALE)
+  0 // terminator
+  --- SQRC Filler Bytes
+  ec11ec11ec11ec // SQRC filler bytes
+  */
+export function rawDataToU8A(rawData) {
+  if (!rawData) {
+    return null;
+  }
+
+  // Strip filler bytes padding at the end
+  if (rawData.substr(-2) === 'ec') {
+    rawData = rawData.substr(0, rawData.length - 2);
+  }
+
+  while (rawData.substr(-4) === 'ec11') {
+    rawData = rawData.substr(0, rawData.length - 4);
+  }
+
+  // Verify that the QR encoding is binary and it's ending with a proper terminator
+  if (rawData.substr(0, 1) !== '4' || rawData.substr(-1) !== '0') {
+    return null;
+  }
+
+  // Strip the encoding indicator and terminator for ease of reading
+  rawData = rawData.substr(1, rawData.length - 2);
+
+  const length8 = parseInt(rawData.substr(0, 2), 16) || 0;
+  const length16 = parseInt(rawData.substr(0, 4), 16) || 0;
+  let length = 0;
+
+  // Strip length prefix
+  if (length8 * 2 + 2 === rawData.length) {
+    rawData = rawData.substr(2);
+    length = length8;
+  } else if (length16 * 2 + 4 === rawData.length) {
+    rawData = rawData.substr(4);
+    length = length16;
+  } else {
+    return null;
+  }
+
+  const bytes = new Uint8Array(length);
+
+  for (let i = 0; i < length; i++) {
+    bytes[i] = parseInt(rawData.substr(i * 2, 2), 16);
+  }
+
+  return bytes;
+}
